[info] Provide floating point info access to player progress bar
This adds a function GetFloat to CGUIInfoManager providing access to
progress bar values as float. The function falls back to GetInt for
other data so that it can be used conveniently in the progress and
slider controls.

The purpose of this change is to allow fine granular progress bars.
Percentage values in the Confluence full screen player progress bar
(720px width) were shown with an inaccuracy of 0-4px because of int
truncation errors. This could lead to a visible gap especially when
used with (chapter) markers.

The values themselves were already calculated as floats and also
used as floats in the gui controls.

+++ xbmc/guilib/GUIProgressControl.cpp
@@ -325,9 +325,9 @@ void CGUIProgressControl::UpdateInfo(const CGUIListItem *item)
   {
     if (m_iInfoCode)
     {
-      int value;
-      if (g_infoManager.GetInt(value, m_iInfoCode, m_parentID, item))
-        m_fPercent = (float)value;
+      float value;
+      if (g_infoManager.GetFloat(value, m_iInfoCode, m_parentID, item))
+        m_fPercent = value;
 
       if (m_fPercent < 0.0f) m_fPercent = 0.0f;
       if (m_fPercent > 100.0f) m_fPercent = 100.0f;
